Continuing Work on Transferring Data Between World and Stages
Continuing work on transferring data between the world and the stages. The transfer of data between the world and the knowledge stage appears to work fine. Still have some details that have to be fleshed out, and stage completion events still need to be added.

+++ Assets/Scripts/Gameplay/World/WorldStage.cs
@@ -46,7 +46,7 @@ namespace RM_EDU
         public int difficulty = 0;
 
         // The score for this stage.
-        public int score = 0;
+        public float score = 0;
 
         // Marker for if the stage is complete.
         public bool complete = false;
